refactor: common-dto-response로 이동 #47
